fix: load glass catalogs in deterministic order
os.listdir returns the files in a directory in arbitrary order, meaning this can differ from machine to machine. KrakenOS will use the first glass with the specified name it finds. For multiple definitions of the same glass name this means the order the catalogs are loaded is very important. Sorting the catalog names will load them in the same order every time.